ambiguous call to overloaded function. Thanks Liquid
https://github.com/patriciogonzalezvivo/ofxFX/issues/4

+++ src/interactive/ofxFlocking.cpp
@@ -321,7 +321,7 @@ void ofxFlocking::allocate(int _width, int _height, int _nBoids){
     
     compileCode();
     
-    resolution = (int) sqrt(_nBoids);
+    resolution = (int) sqrt( (double)_nBoids);
     nBoids = resolution * resolution;
     
     sparkImage.loadImage("spark.png");
